paginacion en las 3 tablas, refactorice el componente contenido, repare el bug con la search bar
