+++ client/src/components/Dashboard/DashboardAdmin.jsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react";
-import style from "./DashboardAdmin.module.css";
+import React, { useEffect } from "react";
 import {
   IconButton,
   Avatar,
@@ -15,31 +14,21 @@ import {
   DrawerContent,
   Text,
   useDisclosure,
-  Button,
   Divider,
 } from "@chakra-ui/react";
 import { FiCompass, FiLink2, FiMenu } from "react-icons/fi";
 import Logo from "../../assets/logo.png";
-import { CircularProgress, CircularProgressLabel } from "@chakra-ui/react";
 import { useDispatch, useSelector } from "react-redux";
 import { useLocation } from "react-router-dom";
 import {
-  adminUser,
-  banComentario,
-  deleteClase,
   getClases,
   getClientes,
   getComentarios,
   getUserByEmail,
-  searchClaseByName,
-  searchClientByEmail,
-  searchComentarioByName,
 } from "../../redux/Actions";
 import { useAuth0 } from "@auth0/auth0-react";
-import { calculoMembresias } from "./calculoMembresias";
-import { banUser } from "../../redux/Actions/index";
-import Swal from "sweetalert2";
-import Search from "./Search/Search";
+
+import Contenido from "./Contenido/Contenido";
 
 const LinkItems = [
   {
@@ -76,11 +65,10 @@ export default function SidebarWithHeader({ children }) {
   const clientes = useSelector((state) => state.clientes);
   const comentarios = useSelector((state) => state.comentarios);
   const clases = useSelector((state) => state.clases);
-  const [client, setClient] = useState("");
+
   const dispatch = useDispatch();
+
   const { pathname } = useLocation();
-  const [clase, setClase] = useState("");
-  const [comentario, setComentario] = useState("");
   useEffect(() => {
     dispatch(getClientes());
     dispatch(getUserByEmail(user?.email));
@@ -88,18 +76,6 @@ export default function SidebarWithHeader({ children }) {
     dispatch(getClases());
   }, [dispatch, user?.email]);
 
-  const handleSubmitClient = (e) => {
-    e.preventDefault();
-    dispatch(searchClientByEmail(client));
-  };
-  const handleSubmitClase = (e) => {
-    e.preventDefault();
-    dispatch(searchClaseByName(clase));
-  };
-  const handleSubmitComentario = (e) => {
-    e.preventDefault();
-    dispatch(searchComentarioByName(comentario));
-  };
   return (
     <Box minH="100vh" bg={useColorModeValue("red.100", "gray.900")}>
       <SidebarContent
@@ -128,15 +104,6 @@ export default function SidebarWithHeader({ children }) {
         comentarios={comentarios}
         clases={clases}
         pathname={pathname}
-        client={client}
-        setClient={setClient}
-        handleSubmitClient={handleSubmitClient}
-        clase={clase}
-        setClase={setClase}
-        handleSubmitClase={handleSubmitClase}
-        comentario={comentario}
-        setComentario={setComentario}
-        handleSubmitComentario={handleSubmitComentario}
       />
       <Box ml={{ base: 0, md: 60 }} p="4">
         {children}
@@ -264,395 +231,5 @@ const MobileNav = ({ admin, onOpen, ...rest }) => {
         </Flex>
       </HStack>
     </Flex>
-  );
-};
-
-const Contenido = ({
-  clientes,
-  comentarios,
-  clases,
-  pathname,
-  client,
-  setClient,
-  handleSubmitClient,
-  clase,
-  setClase,
-  handleSubmitClase,
-  comentario,
-  setComentario,
-  handleSubmitComentario,
-}) => {
-  function refreshPage() {
-    window.location.reload(false);
-  }
-  const dispatch = useDispatch();
-
-  const handleBan = (item) => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You are going to ban/unban the user",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, ban the user!",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        Swal.fire("Banned", "The user has been banned/unbanned", "success");
-        if (item?.isBanned) dispatch(banUser(item?.email, { isBanned: false }));
-        else dispatch(banUser(item?.email, { isBanned: true }));
-      }
-    });
-  };
-
-  const handleAdmin = (item) => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You about to make a user admin or remove his admin privileges",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, do it!",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        Swal.fire("Done", "The user is / is not now an admin", "success");
-        if (item?.isAdmin) dispatch(adminUser(item?.email, { isAdmin: false }));
-        else dispatch(adminUser(item?.email, { isAdmin: true }));
-      }
-    });
-  };
-  const handleBanComent = (item) => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You won't be able to revert this!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, ban it!",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        Swal.fire(
-          "Deleted!",
-          "The comment is not visible in the page and the user has been banned",
-          "success"
-        );
-        if (item?.isBanned) {
-          dispatch(banComentario(item?.id, { isBanned: false }));
-        } else {
-          dispatch(banComentario(item?.id, { isBanned: true }));
-          dispatch(banUser(item?.emailCliente, { isBanned: true }));
-        }
-      }
-    });
-  };
-  const handleDelete = (id) => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You won't be able to revert this!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, delete it!",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        Swal.fire(
-          "Deleted!",
-          "The class has been deleted from the data base",
-          "success"
-        );
-        dispatch(deleteClase(id));
-      }
-    });
-  };
-
-  function paginate(array, tamaño) {
-    var result = [];
-    for (var i = 0; i < array.length; i += tamaño) {
-      result.push(array.slice(i, i + tamaño));
-    }
-    return result;
-  }
-
-  const allClients = useSelector((state) => state.clientes);
-  const pageSize = 3;
-  const [currentPage, setCurrentPage] = useState(1);
-  const pageCount = Math.ceil(allClients.length / pageSize);
-
-  const clientesPaginados = paginate(allClients, pageSize);
-
-  const handlePageChange = (page) => {
-    setCurrentPage(page);
-    console.log(clientesPaginados[page - 1]);
-  };
-
-  return (
-    <Box className={style.container}>
-      <div className={style.titleContainer}>
-        <Text className={style.title}>Admin's Dashboard</Text>
-
-        <Button
-          colorScheme="blue"
-          className={style.refreshButton}
-          onClick={refreshPage}
-        >
-          Refresh Data
-        </Button>
-      </div>
-
-      <Divider
-        marginTop={"50px"}
-        w={"80%"}
-        border={"5px solid white"}
-      ></Divider>
-      {pathname === "/dashboard/stadistics" || pathname === "/dashboard" ? (
-        <div className={style.titleestadisticas}>
-          <div>
-            <Text className={style.clientlist} fontSize="2xl" fontWeight="bold">
-              Membership stadistics:
-            </Text>
-          </div>
-          <div>
-            <Box>
-              <Flex align="center" height="300px">
-                <div className={style.circulo}>
-                  <CircularProgress
-                    value={calculoMembresias(clientes, "Silver")}
-                    color="gray"
-                    size="200px"
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                  >
-                    <CircularProgressLabel>Sil.</CircularProgressLabel>
-                  </CircularProgress>
-                  <CircularProgress
-                    value={calculoMembresias(clientes, "Gold")}
-                    color="gold"
-                    size="200px"
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                  >
-                    <CircularProgressLabel>Gold</CircularProgressLabel>
-                  </CircularProgress>
-                  <CircularProgress
-                    value={calculoMembresias(clientes, "Platinum")}
-                    color="teal"
-                    size="200px"
-                    display="flex"
-                    justifyContent="center"
-                    alignItems="center"
-                  >
-                    <CircularProgressLabel>Plat.</CircularProgressLabel>
-                  </CircularProgress>
-                </div>
-              </Flex>
-            </Box>
-          </div>
-        </div>
-      ) : (
-        <></>
-      )}
-      {pathname === "/dashboard/clients" ? (
-        <>
-          <div className={style.clientSearch}>
-            <Text className={style.clientlist} fontSize="2xl" fontWeight="bold">
-              Clients List:
-            </Text>
-            <Search
-              search={client}
-              setSearch={setClient}
-              handleSubmit={handleSubmitClient}
-            />
-          </div>
-
-          <div className={style.containerListado}>
-            <Box>
-              <table className={style.tabla}>
-                <thead>
-                  <tr>
-                    <th>User</th>
-                    <th>Email</th>
-                    <th>Suscription</th>
-                    <th>Is Banned</th>
-                    <th>Is Admin</th>
-                    <th>Action</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {clientesPaginados[currentPage - 1]?.map((item, index) => (
-                    <tr key={index}>
-                      <td>{item?.nombre}</td>
-                      <td>{item?.email}</td>
-                      <td>{item?.tipoDeSuscripcion}</td>
-                      <td>{item?.isBanned.toString()}</td>
-                      <td>{item?.isAdmin.toString()}</td>
-                      <td className={style.buttonO}>
-                        <Button
-                          colorScheme="red"
-                          onClick={() => handleBan(item)}
-                        >
-                          BAN
-                        </Button>
-                        <Button
-                          colorScheme="green"
-                          onClick={() => handleAdmin(item)}
-                        >
-                          ADMIN
-                        </Button>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-              <div className={style.pagesContainer}>
-                <Button
-                  colorScheme="blackAlpha"
-                  onClick={() =>
-                    handlePageChange(currentPage === 1 ? 1 : currentPage - 1)
-                  }
-                >
-                  {"<"}
-                </Button>
-
-                {Array.from({ length: pageCount }).map((_, index) => (
-                  <Button
-                    colorScheme="blackAlpha"
-                    onClick={() => handlePageChange(index + 1)}
-                  >
-                    {index + 1}
-                  </Button>
-                ))}
-
-                <Button
-                  colorScheme="blackAlpha"
-                  onClick={() =>
-                    handlePageChange(
-                      currentPage === pageCount ? currentPage : currentPage + 1
-                    )
-                  }
-                >
-                  {">"}
-                </Button>
-              </div>
-            </Box>
-          </div>
-        </>
-      ) : (
-        <></>
-      )}
-      {pathname === "/dashboard/classes" ? (
-        <>
-          <div className={style.clientSearch}>
-            <Text className={style.clientlist} fontSize="2xl" fontWeight="bold">
-              Classes:
-            </Text>
-            <Search
-              search={clase}
-              setSearch={setClase}
-              handleSubmit={handleSubmitClase}
-            />
-          </div>
-          <div className={style.containerListado}>
-            <Box>
-              <table className={style.tabla}>
-                <thead>
-                  <tr>
-                    <th>Name</th>
-                    <th>Days</th>
-                    <th>Start time</th>
-                    <th>Delete Class</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {clases?.map((item, index) => (
-                    <tr key={index}>
-                      <td className={style.class}>
-                        {item?.nombre.toUpperCase()}
-                      </td>
-                      <td className={style.days}>{item?.dias.join(" - ")}</td>
-                      <td className={style.schudle}>{item?.horario}</td>
-                      <td className={style.buttonO}>
-                        <Button
-                          colorScheme="red"
-                          onClick={() => handleDelete(item.id)}
-                        >
-                          DELETE
-                        </Button>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </Box>
-          </div>
-        </>
-      ) : (
-        <></>
-      )}
-      {pathname === "/dashboard/comments" ? (
-        <>
-          <div className={style.clientSearch}>
-            <Text
-              className={style.commentsList}
-              fontSize="2xl"
-              fontWeight="bold"
-            >
-              Clients Comments:
-            </Text>
-            <Search
-              search={comentario}
-              setSearch={setComentario}
-              handleSubmit={handleSubmitComentario}
-            />
-          </div>
-          <div className={style.containerListado}>
-            <Box>
-              <table className={style.tabla}>
-                <thead>
-                  <tr>
-                    <th>User</th>
-                    <th>Email</th>
-                    <th>Class</th>
-                    <th>Comment</th>
-                    <th>Is Banned</th>
-                    <th>Action</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {comentarios?.map((item, index) => (
-                    <tr key={index}>
-                      <td className={style.user}>{item?.nombreCliente}</td>
-                      <td className={style.email}>{item?.emailCliente}</td>
-                      <td className={style.class}>{item.nombreClase}</td>
-
-                      <td className={style.comment}>
-                        <textarea disabled={true}>{item?.texto}</textarea>
-                      </td>
-                      <td className={style.isBanned}>
-                        {item?.isBanned.toString()}
-                      </td>
-                      <td className={style.buttonO}>
-                        <Button
-                          onClick={() => handleBanComent(item)}
-                          colorScheme="red"
-                        >
-                          BAN
-                        </Button>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </Box>
-          </div>
-        </>
-      ) : (
-        <></>
-      )}
-    </Box>
   );
 };
